Feat/keypoints

+++ src/Modules/Tours/Explorer.Tours.Core/UseCases/Administration/TourService.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
 using Explorer.BuildingBlocks.Core.UseCases;
 using Explorer.Tours.API.Dtos;
+using Explorer.Tours.API.Dtos.Enums;
 using Explorer.Tours.API.Public.Administration;
 using Explorer.Tours.Core.Domain;
 using Explorer.Tours.Core.Domain.RepositoryInterfaces;
+using System.Net.Sockets;
 namespace Explorer.Tours.Core.UseCases.Administration;
 
 public class TourService : ITourService
@@ -24,7 +26,10 @@ public class TourService : ITourService
         var tour = GetById(id);
         if(tour == null) return false;
 
-        if (tour.Status == TourStatusDTO.Archived) return false;
+        if (tour.Status == TourStatusDto.Archived)
+        {
+            return false;
+        }
 
         Tour? tourToUpdate = _tourRepository.Get(id);
         if (tourToUpdate != null)
@@ -35,14 +40,17 @@ public class TourService : ITourService
         return true;
     }
 
-    public TourDto Create(TourDto tourdto)
+    public TourDto Create(CreateTourDto createTourDto)
     {
-        var result = _tourRepository.Create(_mapper.Map<Tour>(tourdto));
+        var result = _tourRepository.Create(_mapper.Map<Tour>(createTourDto));
         return _mapper.Map<TourDto>(result);
     }
 
-    public void Delete(long id)
+    public void Delete(long id, long authorId)
     {
+        var tour = _tourRepository.Get(id);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't delete someone else's tour");
         _tourRepository.Delete(id);
     }
 
@@ -79,7 +87,7 @@ public class TourService : ITourService
         if (tour == null) return false;
 
 
-        if (tour.Status == TourStatusDTO.Published) return false;
+        if (tour.Status == TourStatusDto.Published) return false;
 
         if (tour.Title.Length <= 0) canPublish = false;
         if (tour.Description.Length <= 0) canPublish = false;
@@ -102,7 +110,7 @@ public class TourService : ITourService
         return canPublish;
     }
 
-    public TourDto Update(long id, TourDto tourDto)
+    public TourDto Update(long id, TourDto tourDto, long authorId)
     {
         var tour = _tourRepository.Get(id);
         if (tour == null)
@@ -110,10 +118,43 @@ public class TourService : ITourService
             throw new KeyNotFoundException($"Tour with id {id} not found.");
         }
 
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't update someone else's tour");
         tour.Update(tourDto.CreatorId, tourDto.Title, tourDto.Description, tourDto.Difficulty,
             tourDto.Tags, (TourStatus)tourDto.Status, tourDto.Price);
 
         var result = _tourRepository.Update(tour);
         return _mapper.Map<TourDto>(result);
+    }
+
+    public KeypointDto AddKeypoint(long tourId, KeypointDto keypointDto, long authorId)
+    {
+        var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't add keypoint to someone else's tour");
+        var keypoint = tour.AddKeypoint(_mapper.Map<Keypoint>(keypointDto));
+        _tourRepository.Update(tour);
+
+        return _mapper.Map<KeypointDto>(keypoint);
+    }
+
+    public KeypointDto UpdateKeypoint(long tourId, KeypointDto keypointDto, long authorId)
+    {
+        var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't update keypoint from someone else's tour");
+        var keypoint = tour.UpdateKeypoint(_mapper.Map<Keypoint>(keypointDto));
+        _tourRepository.Update(tour);
+
+        return _mapper.Map<KeypointDto>(keypoint);
+    }
+
+    public void DeleteKeypoint(long tourId, long keypointId, long authorId)
+    {
+        var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't delete keypoint from someone else's tour");
+        tour.DeleteKeypoint(keypointId);
+        var result = _tourRepository.Update(tour);
     }
 }
